# binary/reader: Return to pool after ReadValue